clearified command... cloc at level of subDir
command for the program does not allow cloc to perform on directory and all subdirectories. 
the command instead allows cloc to be performed on all subdirectories at a particular level. 
number of nests determines which level is accessed.
0 nest  = level 1.
1 nest  = level 2.
2 nest  = level 3.
3 nest  = level 4.
...

this command can be automated to run and then nest and then run. In oder to perform cloc on directory and all subdirectories.
but is not currently shown on the readme. However performing cloc on a particular level is useful. 
See the comment I made on issue 58.