Expand use of cloc;
apply recursive cloc functionality on everything inside but not below "Projects" Folder.
In "Projects" Folder, both files are individual_projects and folders are individual_projects.

This was useful when I was reading a programming book.
Every project was not in a project folder. Sometimes a project was an individual_file.

an example of what I am writing about:
https://github.com/miketestgit01/Art_of_Exploitation_SC/tree/master/Other_Code