Python wrapper script "clocs"
"clocs" is a wrapper script of "cloc", written in python. It can count lines of code within a project directory and/or its subdirectories (the subdirectories' name are automatically acquired by the code, at line 25-28. In my project, "cloc" is in (project)/tools/third-party/, and "clocs" is in (project)/tools/.
The file excluded-files-list required by "cloc --exclude-list-file" should be in (project)/tools/third-party/, as indicated at line 31.
Usage: 
"clocs -h" for helper, 
"clocs -c" to clean all reports (named "stat.txt"), 
"clocs -a" to generate "cloc" reports in the project directory and all subdirectories, 
"clocs ." or "clocs" to generate a report directly in the project directory.

Any feedback is welcome:)