Add support for gencat NLS Catalogue source files
The gencat utility is a part of the X/Open Portability Guide
and in freebsd the standard suffix .msg is used for the input
files. Set the scale factor to 1.50 but am unsuer if this is
correct, its same used by the 'PO File' translation files. Add
a simple test example based on a modified freebsd file.