rm_comments_in_strings: fix for edge case
Fix to correctly handle a situation in rm_comments_in_strings where
a line ends with a multiline string marker that starts a new multiline
string. In such cases, rm_comments_in_strings should just copy this part
of the line without modifications.

Also add some comments and improve indentation to this part of the
code for better readability.