Fix incorrect parsing of config file
This fixes an issue with how the verbose flag is matched in config files. In the current release, it checks for "verbose" or "v" anywhere in the line, instead of only at the beginning of the line. This causes a config like "force-lang Objective-C,h" to match the verbose check instead (because of the v in Objective), if verbose has not already been specified. This changes that check to require v/verbose to be at the start of the line.