Add support for AnsProlog (Answer Set Programming)
[Answer Set Programming](https://en.wikipedia.org/wiki/Answer_set_programming) (ASP) is
a declarative programming paradigm oriented towards combinatorial search problems. Most
ASP solvers accept a language called AnsProlog ("Answer Set Programming in Logic"),
which is enough like Prolog for its lines to be counted the same way, but it typically
uses the `.lp` ("logic program") suffix.

ASP is different enough from Prolog that it deserves to be reported separately -- this
patch will cause ASP code to be reported as AnsProlog, while regular Prolog programs
will be reported as Prolog, so you can tell them apart if they're in the same
repository.

- [x] Add filters for AnsProlog
- [x] Add `.lp` suffix for AnsProlog
- [x] Add scale factor for AnsProlog
- [x] Update `README.md`