add BitBake to .conf multi-language disambiguation
BitBake uses conf files for layers, local and machine specific
assignments. If we don't detect a BitBake marker, we leave it as unknown
(the previous behavior). The detector function uses the same markers as
PHP/Pascal/Fortran/Pawn/BitBake with a few added variables specific to
layer.conf: BBFILE* and LAYER*.